+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.2.12 (2022-12-06)
+* Improved resolution subclass check
+
 ## v0.2.11 (2022-11-28)
 * Added default fallback to resolve handler